fix: fix CR comment
