+++ src/entries/shared/types/storages/config.ts
@@ -217,6 +217,8 @@ export interface IConfigPiniaStorageSchema {
 
     // 是否启用快速站点过滤功能
     quickSiteFilter: boolean;
+    // 是否在搜索框中展示热门推荐按钮
+    showHotRecommendations: boolean;
   };
 
   // 配置同样在 searchEntity 页面（偷懒下）
