+++ src/locales/en.json
@@ -652,6 +652,7 @@
       "forceImdbIdMatchFilter": "When searching by IMDb, filter out results whose IMDb ID returned by the site does not match",
       "autoDetectOfficialGroupFromTitle": "Auto Detect Official Tag From Torrent Title",
       "quickSiteFilter": "Quick Site Filter In Search Page",
+      "showHotRecommendations": "Show hot recommendations button in the search box",
       "mediaServerSearchConfig": "Media Server Search Configuration",
       "mediaQueueConcurrency": "Request Queue Concurrency",
       "mediaSearchLimit": "Number of search results per request",
