+++ src/locales/zh_CN.json
@@ -652,6 +652,7 @@
       "forceImdbIdMatchFilter": "使用 IMDb 搜索时，在搜索结果中滤去站点返回的IMDb号不匹配的结果",
       "autoDetectOfficialGroupFromTitle": "自动识别种子标题中的官方组信息，并形成标签",
       "quickSiteFilter": "直接显示站点筛选栏",
+      "showHotRecommendations": "在搜索框中显示热门推荐按钮",
       "mediaServerSearchConfig": "媒体服务器搜索配置",
       "mediaQueueConcurrency": "请求队列长度",
       "mediaSearchLimit": "单次搜索结果数量",
